requirements.txt
ansible ~=2.10.4
boto3 ~=1.16.47
Fabric3 ~=1.14
invoke ~=1.5.0
python-dotenv ~=0.15.0
sshtunnel ~=0.4.0
